Locker constructors, init method + framework and docstrings for all
First construcor takes in path to the locker and loads it. If boolean
isnew is true, creates a new locker at the path. Second takes in a
windows size int input that sets the rabin finger print params such that
average chunk size is close to the input size if creating a new locker.
Init methods sets up a new locker, sets up all the required metadata and
locker dirs and saves them to disk.

Also added more refined framework and doc strings for all the rest of
the member methods.